Add SqlDelight to JS!! Adapt logic for it
Still not working

+++ shared/interop-sample/data/src/commonMain/kotlin/interopSample/cloud/dtos/WeatherDTO.kt
@@ -13,7 +13,7 @@ internal data class WeatherDTO(
     @Serializable
     internal data class CurrentWeatherDTO(
         @SerialName("temperature")
-        val temperature: Float
+        val temperature: Double
     )
 
     @Serializable
